feat: fix unit tests after merge main Signed-off-by: Cristian G <cristian.castro@iohk.io>
