+++ edge-agent-sdk/src/commonMain/kotlin/org/hyperledger/identus/walletsdk/domain/models/JWTCredentialPayload.kt
@@ -3,13 +3,21 @@ package org.hyperledger.identus.walletsdk.domain.models
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.SerializationException
 import kotlinx.serialization.builtins.MapSerializer
 import kotlinx.serialization.builtins.serializer
+import kotlinx.serialization.descriptors.SerialDescriptor
+import kotlinx.serialization.descriptors.buildClassSerialDescriptor
+import kotlinx.serialization.descriptors.element
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
+import kotlinx.serialization.encoding.encodeStructure
+import kotlinx.serialization.json.JsonDecoder
 import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.JsonPrimitive
+import kotlinx.serialization.json.int
+import kotlinx.serialization.json.jsonObject
 import kotlinx.serialization.json.jsonPrimitive
 
 interface JWTPayload {
@@ -61,7 +69,7 @@ data class JWTVerifiableCredential @JvmOverloads constructor(
     val credentialSchema: VerifiableCredentialTypeContainer? = null,
     @Serializable(with = MapStringAnyToStringSerializer::class)
     val credentialSubject: Map<String, String>,
-    val credentialStatus: VerifiableCredentialTypeContainer? = null,
+    val credentialStatus: CredentialStatus? = null,
     val refreshService: VerifiableCredentialTypeContainer? = null,
     val evidence: VerifiableCredentialTypeContainer? = null,
     val termsOfUse: VerifiableCredentialTypeContainer? = null
@@ -106,6 +114,87 @@ data class JWTVerifiableCredential @JvmOverloads constructor(
         result = 31 * result + (evidence?.hashCode() ?: 0)
         result = 31 * result + (termsOfUse?.hashCode() ?: 0)
         return result
+    }
+
+    @Serializable(with = CredentialStatusSerializer::class)
+    data class CredentialStatus(
+        val id: String,
+        val type: CredentialStatusListType,
+        val statusPurpose: CredentialStatusPurpose,
+        val statusListIndex: Int,
+        val statusListCredential: String
+    )
+
+    enum class CredentialStatusListType(val type: String) {
+        // The naming does not follow the proper kotlin format, but it is required to work around a
+        // situation when serializing an object containing an instance of this enum from nimbus library
+        // using java serialization instead of kotlin x.
+        StatusList2021Entry("StatusList2021Entry");
+
+        companion object {
+            fun fromString(type: String): CredentialStatusListType {
+                return entries.first { it.type.lowercase() == type.lowercase() }
+            }
+        }
+    }
+
+    enum class CredentialStatusPurpose(val purpose: String) {
+        REVOCATION("revocation"),
+        SUSPENSION("suspension");
+
+        companion object {
+            fun fromString(purpose: String): CredentialStatusPurpose {
+                return entries.first { it.purpose.lowercase() == purpose.lowercase() }
+            }
+        }
+    }
+
+    object CredentialStatusSerializer : KSerializer<CredentialStatus> {
+        override val descriptor: SerialDescriptor = buildClassSerialDescriptor("credentialStatus") {
+            element<String>("id")
+            element<String>("type")
+            element<String>("statusPurpose")
+            element<Int>("statusListIndex")
+            element<String>("statusListCredential")
+        }
+
+        override fun serialize(encoder: Encoder, value: CredentialStatus) {
+            encoder.encodeStructure(descriptor) {
+                encodeStringElement(descriptor, 0, value.id)
+                encodeStringElement(descriptor, 1, value.type.type)
+                encodeStringElement(descriptor, 2, value.statusPurpose.purpose)
+                encodeIntElement(descriptor, 3, value.statusListIndex)
+                encodeStringElement(descriptor, 4, value.statusListCredential)
+            }
+        }
+
+        override fun deserialize(decoder: Decoder): CredentialStatus {
+            require(decoder is JsonDecoder)
+            val jsonObject = decoder.decodeJsonElement().jsonObject
+
+            val id = jsonObject["id"]?.jsonPrimitive?.content ?: throw SerializationException("Missing 'id'")
+            val type = CredentialStatusListType.fromString(
+                jsonObject["type"]?.jsonPrimitive?.content ?: throw SerializationException("Missing 'type'")
+            )
+
+            val statusPurpose = CredentialStatusPurpose.fromString(
+                jsonObject["statusPurpose"]?.jsonPrimitive?.content
+                    ?: throw SerializationException("Missing 'statusPurpose'")
+            )
+            val statusListIndex = jsonObject["statusListIndex"]?.jsonPrimitive?.int ?: throw SerializationException(
+                "Missing 'statusListIndex'"
+            )
+            val statusListCredential = jsonObject["statusListCredential"]?.jsonPrimitive?.content
+                ?: throw SerializationException("Missing 'statusListCredential'")
+
+            return CredentialStatus(
+                id = id,
+                type = type,
+                statusPurpose = statusPurpose,
+                statusListIndex = statusListIndex,
+                statusListCredential = statusListCredential
+            )
+        }
     }
 }
